Update binary_numbers.py
This solution and exercise are appropriate for practicing Enqueue and Dequeue operations of Queue Data Structure but this solution for this exercise is storing an additional 11 binary numbers in the queue and using up extra space in the memory which makes it less efficient. 
Maybe you can use the Clear() method to clear the queue or Count() method to stop the for loop if the queue already has 10 elements and then dequeue the queue.